requirements.txt
# Framework core
fastapi==0.112.1
uvicorn[standard]==0.30.6
sqlmodel==0.0.22

# Database (psycopg2 compila desde fuente, compatible con cualquier Python)
# Nota: La primera compilación tomará más tiempo
psycopg2>=2.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Environment
python-dotenv==1.0.1
tzdata==2024.2

# HTTP Client (mantenido por compatibilidad, ya no se usa activamente)
aiohttp==3.9.0

# Data validation
pydantic==2.8.2
pydantic-core==2.20.1

# Email (si se usa en el futuro)
email-validator==2.2.0

# Utils (dependencias transitivas de uvicorn, necesarias)
annotated-types==0.7.0
anyio==4.4.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.7
cloudinary==1.42.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
frozenlist==1.7.0
h11==0.14.0
idna==3.7
multidict==6.6.3
propcache==0.3.2
pyasn1==0.6.1
PyJWT==2.9.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.20.1
# uvicorn[standard] dependencies
httptools==0.7.1
pyyaml==6.0.3
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
# cryptography dependencies
cffi==2.0.0
pycparser==3.0
# aiohttp dependencies
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
# passlib[bcrypt] dependency (compatible con passlib 1.7.4)
bcrypt==4.2.1
# email-validator dependencies
typing-extensions==4.12.2
